+++ test/nextgen/operators/matmul/matmul/matmul_wrapper_registry.cpp
@@ -341,4 +341,27 @@ std::unique_ptr<KernelWrapper<MatMulShape>> create_matmul_clamp_qai8_qai8_qsi4cx
     return create_matmul_clamp_qai8_qsi4(false);
 }
 
+std::unique_ptr<KernelWrapper<MatMulShape>>
+create_matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa() {
+    MatMulUkerOutputStageConfig output_stage_config{};
+    output_stage_config.scale_bias = {MatMulUkerStageParameterLayout::GLOBAL};
+    output_stage_config.scale_bias_global_slot = MatMulSlot::DST_QZP;
+    output_stage_config.extra_ref_inputs = {
+        MatMulSlot::LHS_QDATA,         MatMulSlot::LHS_QSCALE,   MatMulSlot::LHS_QZP,
+        MatMulSlot::RHS_T_QDATA,       MatMulSlot::RHS_T_QSCALE, MatMulSlot::ACC_BIAS_N_QDATA,
+        MatMulSlot::ACC_BIAS_N_QSCALE, MatMulSlot::DST_QSCALE,   MatMulSlot::DST_QZP};
+
+    return std::make_unique<MatMulUkerApiWrapper>(
+        "matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa",
+        kai_matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa(), MatMulSlot::LHS_PACKED,
+        make_poly<Block2dRowFormat>(
+            4 * get_sme_vector_scale(), 4, 4, false, DataType::I8, std::array<DataType, 0>{},
+            std::array<DataType, 0>{}),
+        make_poly<Block2dRowFormat>(
+            4 * get_sme_vector_scale(), 4, 4, false, DataType::I8, std::array{DataType::I32},
+            std::array{DataType::FP32}),
+        make_poly<PlainFormat>(DataType::I8), DataType::I32, MatMulUkerClampConfig::optional(DataType::I32),
+        MatMulUkerApiBiasDeliveryStage::PACK_RHS, output_stage_config);
+}
+
 }  // namespace kai::test