+++ test/nextgen/operators/matmul/matmul_operator.cpp
@@ -50,7 +50,7 @@ const MatMulBiasModeSet acc_bias_per_m_per_n_scale_bias_per_n{
 }  // namespace
 
 Span<const MatMulOperator> get_available_matmul_operators() {
-    static std::array<MatMulOperator, 25> operators;
+    static std::array<MatMulOperator, 26> operators;
 
     // matmul_clamp_f32_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa
     operators[0].name = "matmul_clamp_f32_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa";
@@ -507,36 +507,41 @@ Span<const MatMulOperator> get_available_matmul_operators() {
     operators[18].pack_rhs = create_matmul_pack_rhs_nxk_qsi4cxp8vsx4sf32bi32_qsi4cx_f32_i32_sme();
     operators[18].matmul = create_matmul_clamp_qai8_qai8_qsi4cxp8vsx4sf32bi32_1x64vs_sme2_dot();
 
-    // kai_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa - KxN RHS pack
-    operators[19].name = "matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa_rhs_kxn";
+    // kai_matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa
+    operators[19].name = "matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa";
 
-    operators[19].is_cpu_supported = cpu_has_sme;
+    operators[19].is_cpu_supported = cpu_has_sme_mop4;
     operators[19].is_shape_suitable = all_true<  //
-        is_shape_suitable_lhs_x8p4vsx4_x8_sme,  //
-        is_shape_suitable_rhs_kxn_x8p4vsx4_x8_sme>;
-    operators[19].supported_bias_mode_sets = {acc_bias_per_m_per_n};
-    operators[19].clamp_mode = MatMulClampMode::UNSUPPORTED;
-    operators[19].lhs_quant = std::nullopt;
-    operators[19].rhs_quant = std::nullopt;
-    operators[19].bias_quant = std::nullopt;
-    operators[19].lhs_dtype = DataType::U8;
-    operators[19].rhs_dtype = DataType::U8;
-    operators[19].bias_dtype = DataType::I32;
+        is_shape_suitable_lhs_x8p4vsx4_x8_sme,   //
+        is_shape_suitable_rhs_kxn_qsi8cxp4vsx4bi32sf32_qsi8_i32_f32_sme>;
+    operators[19].supported_bias_mode_sets = {acc_bias_per_n};
+    operators[19].clamp_mode = MatMulClampMode::OPTIONAL;
+    operators[19].lhs_quant = std::make_unique<AsymmLinearQuantizer>(
+        DataType::I8, DataType::FP32, DataType::I32, RoundMode::TIE_AWAY, RoundMode::CURRENT, 0, 0);
+    operators[19].rhs_quant =
+        std::make_unique<SymmLinearQuantizer>(DataType::I8, DataType::FP32, RoundMode::CURRENT, 1, 0);
+    operators[19].bias_quant =
+        std::make_unique<SymmLinearQuantizer>(DataType::I32, DataType::FP32, RoundMode::CURRENT, 1, 1);
+    operators[19].dst_quant = std::make_unique<AsymmLinearQuantizer>(
+        DataType::I8, DataType::FP32, DataType::I32, RoundMode::TIE_AWAY, RoundMode::CURRENT, 0, 0);
+    operators[19].bias_quant_info_source = MatMulBiasQuantInfoSource::STATIC_FROM_INPUT_AND_OUTPUT_QUANT;
+    operators[19].lhs_dtype = DataType::FP32;
+    operators[19].rhs_dtype = DataType::FP32;
+    operators[19].bias_dtype = DataType::FP32;
     operators[19].acc_dtype = DataType::I32;
-    operators[19].dst_dtype = DataType::I32;
-    operators[19].ref_dtype = DataType::I32;
+    operators[19].dst_dtype = DataType::I8;
 
-    operators[19].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme();
-    operators[19].pack_rhs = create_matmul_pack_rhs_kxn_x8p4vsx4_x8_sme();
-    operators[19].matmul = create_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa();
+    operators[19].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_i8_sme();
+    operators[19].pack_rhs = create_matmul_pack_rhs_kxn_qsi8cxp4vsx4bi32sf32_qsi8_i32_f32_sme();
+    operators[19].matmul = create_matmul_clamp_qai8_qai8p4vsx4_qsi8cxp4vsx4sf32bi32_8vsx8vs_sme2p1_mop4_mopa();
 
-    // kai_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa - NxK RHS pack
-    operators[20].name = "matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa_rhs_nxk";
+    // kai_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa - KxN RHS pack
+    operators[20].name = "matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa_rhs_kxn";
 
     operators[20].is_cpu_supported = cpu_has_sme;
     operators[20].is_shape_suitable = all_true<  //
         is_shape_suitable_lhs_x8p4vsx4_x8_sme,  //
-        is_shape_suitable_rhs_nxk_x8p4vsx4_x8_sme>;
+        is_shape_suitable_rhs_kxn_x8p4vsx4_x8_sme>;
     operators[20].supported_bias_mode_sets = {acc_bias_per_m_per_n};
     operators[20].clamp_mode = MatMulClampMode::UNSUPPORTED;
     operators[20].lhs_quant = std::nullopt;
@@ -550,18 +555,18 @@ Span<const MatMulOperator> get_available_matmul_operators() {
     operators[20].ref_dtype = DataType::I32;
 
     operators[20].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme();
-    operators[20].pack_rhs = create_matmul_pack_rhs_nxk_x8p4vsx4_x8_sme();
+    operators[20].pack_rhs = create_matmul_pack_rhs_kxn_x8p4vsx4_x8_sme();
     operators[20].matmul = create_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa();
 
-    // kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa - KxN RHS pack
-    operators[21].name = "matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa_rhs_kxn";
+    // kai_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa - NxK RHS pack
+    operators[21].name = "matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa_rhs_nxk";
 
     operators[21].is_cpu_supported = cpu_has_sme;
     operators[21].is_shape_suitable = all_true<  //
         is_shape_suitable_lhs_x8p4vsx4_x8_sme,  //
-        is_shape_suitable_rhs_kxn_x8p4vsx4_x8_sme>;
-    operators[21].supported_bias_mode_sets = {acc_bias_per_m_per_n_scale_bias_per_n};
-    operators[21].clamp_mode = MatMulClampMode::OPTIONAL;
+        is_shape_suitable_rhs_nxk_x8p4vsx4_x8_sme>;
+    operators[21].supported_bias_mode_sets = {acc_bias_per_m_per_n};
+    operators[21].clamp_mode = MatMulClampMode::UNSUPPORTED;
     operators[21].lhs_quant = std::nullopt;
     operators[21].rhs_quant = std::nullopt;
     operators[21].bias_quant = std::nullopt;
@@ -569,20 +574,20 @@ Span<const MatMulOperator> get_available_matmul_operators() {
     operators[21].rhs_dtype = DataType::U8;
     operators[21].bias_dtype = DataType::I32;
     operators[21].acc_dtype = DataType::I32;
-    operators[21].dst_dtype = DataType::FP32;
+    operators[21].dst_dtype = DataType::I32;
     operators[21].ref_dtype = DataType::I32;
 
     operators[21].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme();
-    operators[21].pack_rhs = create_matmul_pack_rhs_kxn_x8p4vsx4_x8_sme();
-    operators[21].matmul = create_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa();
+    operators[21].pack_rhs = create_matmul_pack_rhs_nxk_x8p4vsx4_x8_sme();
+    operators[21].matmul = create_matmul_i32_u8p4vsx4_u8p4vsx4_i32_i32_8vsx8vs_qmx_mopa();
 
-    // kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa - NxK RHS pack
-    operators[22].name = "matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa_rhs_nxk";
+    // kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa - KxN RHS pack
+    operators[22].name = "matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa_rhs_kxn";
 
     operators[22].is_cpu_supported = cpu_has_sme;
     operators[22].is_shape_suitable = all_true<  //
-        is_shape_suitable_lhs_x8p4vsx4_x8_sme,   //
-        is_shape_suitable_rhs_nxk_x8p4vsx4_x8_sme>;
+        is_shape_suitable_lhs_x8p4vsx4_x8_sme,  //
+        is_shape_suitable_rhs_kxn_x8p4vsx4_x8_sme>;
     operators[22].supported_bias_mode_sets = {acc_bias_per_m_per_n_scale_bias_per_n};
     operators[22].clamp_mode = MatMulClampMode::OPTIONAL;
     operators[22].lhs_quant = std::nullopt;
@@ -596,40 +601,40 @@ Span<const MatMulOperator> get_available_matmul_operators() {
     operators[22].ref_dtype = DataType::I32;
 
     operators[22].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme();
-    operators[22].pack_rhs = create_matmul_pack_rhs_nxk_x8p4vsx4_x8_sme();
+    operators[22].pack_rhs = create_matmul_pack_rhs_kxn_x8p4vsx4_x8_sme();
     operators[22].matmul = create_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa();
 
-    // kai_matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla - non-transposed RHS.
-    operators[23].name = "matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla";
+    // kai_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa - NxK RHS pack
+    operators[23].name = "matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa_rhs_nxk";
 
     operators[23].is_cpu_supported = cpu_has_sme;
-    operators[23].is_shape_suitable = all_true<   //
-        is_shape_suitable_lhs_vector,             //
-        is_shape_suitable_lhs_x32p4vsx1_x32_sme,  //
-        is_shape_suitable_rhs_kxn_x32p4vsx1bx32_x32_x32_sme>;
-    operators[23].supported_bias_mode_sets = {acc_bias_per_n};
+    operators[23].is_shape_suitable = all_true<  //
+        is_shape_suitable_lhs_x8p4vsx4_x8_sme,   //
+        is_shape_suitable_rhs_nxk_x8p4vsx4_x8_sme>;
+    operators[23].supported_bias_mode_sets = {acc_bias_per_m_per_n_scale_bias_per_n};
     operators[23].clamp_mode = MatMulClampMode::OPTIONAL;
     operators[23].lhs_quant = std::nullopt;
     operators[23].rhs_quant = std::nullopt;
     operators[23].bias_quant = std::nullopt;
-    operators[23].lhs_dtype = DataType::FP32;
-    operators[23].rhs_dtype = DataType::FP32;
-    operators[23].bias_dtype = DataType::FP32;
-    operators[23].acc_dtype = DataType::FP32;
+    operators[23].lhs_dtype = DataType::U8;
+    operators[23].rhs_dtype = DataType::U8;
+    operators[23].bias_dtype = DataType::I32;
+    operators[23].acc_dtype = DataType::I32;
     operators[23].dst_dtype = DataType::FP32;
+    operators[23].ref_dtype = DataType::I32;
 
-    operators[23].pack_lhs = std::nullopt;
-    operators[23].pack_rhs = create_matmul_pack_rhs_kxn_x32p4vsx1bx32_x32_x32_sme();
-    operators[23].matmul = create_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla();
+    operators[23].pack_lhs = create_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme();
+    operators[23].pack_rhs = create_matmul_pack_rhs_nxk_x8p4vsx4_x8_sme();
+    operators[23].matmul = create_matmul_clamp_f32_u8p4vsx4_u8p4vsx4_i32_i32_f32_f32_8vsx8vs_qmx_mopa();
 
-    // kai_matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla - transposed RHS.
-    operators[24].name = "matmul_clamp_t_f32_f32_f32p4vsx1b_1x32vs_qmx_mla";
+    // kai_matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla - non-transposed RHS.
+    operators[24].name = "matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla";
 
     operators[24].is_cpu_supported = cpu_has_sme;
     operators[24].is_shape_suitable = all_true<   //
         is_shape_suitable_lhs_vector,             //
         is_shape_suitable_lhs_x32p4vsx1_x32_sme,  //
-        is_shape_suitable_rhs_nxk_x32p4vsx1bx32_x32_x32_sme>;
+        is_shape_suitable_rhs_kxn_x32p4vsx1bx32_x32_x32_sme>;
     operators[24].supported_bias_mode_sets = {acc_bias_per_n};
     operators[24].clamp_mode = MatMulClampMode::OPTIONAL;
     operators[24].lhs_quant = std::nullopt;
@@ -642,8 +647,31 @@ Span<const MatMulOperator> get_available_matmul_operators() {
     operators[24].dst_dtype = DataType::FP32;
 
     operators[24].pack_lhs = std::nullopt;
-    operators[24].pack_rhs = create_matmul_pack_rhs_nxk_x32p4vsx1bx32_x32_x32_sme();
+    operators[24].pack_rhs = create_matmul_pack_rhs_kxn_x32p4vsx1bx32_x32_x32_sme();
     operators[24].matmul = create_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla();
+
+    // kai_matmul_clamp_f32_f32_f32p4vsx1b_1x32vs_qmx_mla - transposed RHS.
+    operators[25].name = "matmul_clamp_t_f32_f32_f32p4vsx1b_1x32vs_qmx_mla";
+
+    operators[25].is_cpu_supported = cpu_has_sme;
+    operators[25].is_shape_suitable = all_true<   //
+        is_shape_suitable_lhs_vector,             //
+        is_shape_suitable_lhs_x32p4vsx1_x32_sme,  //
+        is_shape_suitable_rhs_nxk_x32p4vsx1bx32_x32_x32_sme>;
+    operators[25].supported_bias_mode_sets = {acc_bias_per_n};
+    operators[25].clamp_mode = MatMulClampMode::OPTIONAL;
+    operators[25].lhs_quant = std::nullopt;
+    operators[25].rhs_quant = std::nullopt;
+    operators[25].bias_quant = std::nullopt;
+    operators[25].lhs_dtype = DataType::FP32;
+    operators[25].rhs_dtype = DataType::FP32;
+    operators[25].bias_dtype = DataType::FP32;
+    operators[25].acc_dtype = DataType::FP32;
+    operators[25].dst_dtype = DataType::FP32;
+
+    operators[25].pack_lhs = std::nullopt;
+    operators[25].pack_rhs = create_matmul_pack_rhs_nxk_x32p4vsx1bx32_x32_x32_sme();
+    operators[25].matmul = create_matmul_clamp_f32_f32_f32p4vsx1bf32_1x32vs_qmx_mla();
 
     return operators;
 }